Added custom logger example.
This example disables Boost.Log and uses simple std::log.

Updated document.
