+++ doc/modules/ROOT/nav.adoc
@@ -38,3 +38,4 @@
 ** xref:tool/broker.adoc[]
 ** xref:tool/client_cli.adoc[]
 ** xref:tool/bench.adoc[]
+* link:http://github.com/redboltz/async_mqtt[Repository]
