+++ example/CMakeLists.txt
@@ -4,6 +4,7 @@ list(APPEND exec_PROGRAMS
     ep_slcoro_mqtt_client.cpp
     cl_cpp17_mqtt_pub.cpp
     cl_cpp17_mqtt_sub.cpp
+    custom_logger.cpp
 )
 
 if("cxx_std_20" IN_LIST CMAKE_CXX_COMPILE_FEATURES)
